[scanner] test: split oversized hooks test files (batch D)
Splits 7 oversized test files (each 1088-1197 lines) into smaller
focused files grouped by concern:

- useCachedData.fetchers.test.ts (1197 lines) →
  .fetchers.branch-coverage.test.ts + .fetchers.patterns.test.ts

- useStackDiscovery.test.ts (1172 lines) →
  .discovery.test.ts + .advanced.test.ts

- useCachedData.progressive-gpu.test.ts (1159 lines) →
  .progressive-gpu.clusters.test.ts + .progressive-gpu.hooks.test.ts

- useCachedLLMd.test.ts (1157 lines) →
  .hooks.test.ts + .fetchers.test.ts

- useMarketplace.test.ts (1155 lines) →
  .fetch-filter.test.ts + .stats-install.test.ts + .author.test.ts

- useSearchIndex.categories.test.ts (1095 lines) →
  .categories.base.test.ts + .categories.dashboard.test.ts + .categories.edgecases.test.ts

- useMCS.test.ts (1088 lines) →
  .status-exports-imports.test.ts + .export-import-additional.test.ts
