+++ web/src/hooks/__tests__/useMarketplace.stats-install.test.ts
@@ -184,393 +184,7 @@ describe('useMarketplace', () => {
 
   // ──────────────────────── Basic shape ────────────────────────
 
-  it('returns the expected hook shape', () => {
-    const { result } = renderHook(() => useMarketplace())
-    expect(result.current).toHaveProperty('items')
-    expect(result.current).toHaveProperty('allItems')
-    expect(result.current).toHaveProperty('allTags')
-    expect(result.current).toHaveProperty('typeCounts')
-    expect(result.current).toHaveProperty('cncfStats')
-    expect(result.current).toHaveProperty('cncfCategories')
-    expect(result.current).toHaveProperty('isLoading')
-    expect(result.current).toHaveProperty('error')
-    expect(result.current).toHaveProperty('searchQuery')
-    expect(result.current).toHaveProperty('selectedTag')
-    expect(result.current).toHaveProperty('selectedType')
-    expect(result.current).toHaveProperty('showHelpWanted')
-    expect(result.current).toHaveProperty('installItem')
-    expect(result.current).toHaveProperty('removeItem')
-    expect(result.current).toHaveProperty('isInstalled')
-    expect(result.current).toHaveProperty('getInstalledDashboardId')
-    expect(result.current).toHaveProperty('refresh')
-  })
-
-  // ──────────────────────── Cache behaviour ────────────────────────
-
-  it('loads items via fetch on mount', async () => {
-    const items = [makeItem({ id: 'cached-1', name: 'Cached Dashboard' })]
-    seedCache(items)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems.length).toBe(1)
-    expect(result.current.allItems[0].name).toBe('Cached Dashboard')
-    expect(globalThis.fetch).toHaveBeenCalled()
-  })
-
-  it('fetches fresh data from network on mount', async () => {
-    const freshItems = [makeItem({ id: 'fresh-1', name: 'Fresh' })]
-    seedCache(freshItems)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems[0].id).toBe('fresh-1')
-  })
-
-  it('recovers and loads items after a failed initial fetch', async () => {
-    vi.mocked(globalThis.fetch).mockRejectedValueOnce(new Error('transient'))
-    const items = [makeItem({ id: 'recovered' })]
-    vi.mocked(globalThis.fetch).mockResolvedValueOnce({
-      ok: true,
-      json: () => Promise.resolve(makeRegistry(items)),
-    } as Response)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.error).toBe('transient')
-
-    act(() => { result.current.refresh() })
-    await waitFor(() => {
-      expect(result.current.allItems.length).toBe(1)
-    })
-    expect(result.current.allItems[0].id).toBe('recovered')
-  })
-
-
-  // ──────────────────────── Network fetch ────────────────────────
-
-  it('fetches registry from network on first load', async () => {
-    const items = [makeItem({ id: 'net-1' })]
-    vi.mocked(globalThis.fetch).mockResolvedValueOnce({
-      ok: true,
-      json: () => Promise.resolve(makeRegistry(items)),
-    } as Response)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems.length).toBe(1)
-    expect(result.current.allItems[0].id).toBe('net-1')
-  })
-
-  it('sets error on HTTP failure', async () => {
-    vi.mocked(globalThis.fetch).mockResolvedValueOnce({
-      ok: false,
-      status: 500,
-    } as Response)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.error).toContain('500')
-    expect(result.current.allItems).toEqual([])
-  })
-
-  it('sets error on network failure', async () => {
-    vi.mocked(globalThis.fetch).mockRejectedValueOnce(new Error('Network down'))
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.error).toBe('Network down')
-    expect(result.current.allItems).toEqual([])
-  })
-
-  it('sets generic error for non-Error throws', async () => {
-    vi.mocked(globalThis.fetch).mockRejectedValueOnce('string error')
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.error).toBe('Failed to load marketplace')
-  })
-
-
-  // ──────────────────────── Refresh (skipCache) ────────────────────────
-
-  it('refresh() re-fetches and updates items', async () => {
-    vi.mocked(globalThis.fetch).mockResolvedValueOnce({
-      ok: true,
-      json: () => Promise.resolve(makeRegistry([makeItem({ id: 'old' })])),
-    } as Response)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems[0].id).toBe('old')
-
-    vi.mocked(globalThis.fetch).mockResolvedValueOnce({
-      ok: true,
-      json: () => Promise.resolve(makeRegistry([makeItem({ id: 'refreshed' })])),
-    } as Response)
-
-    act(() => { result.current.refresh() })
-
-    await waitFor(() => {
-      expect(result.current.allItems[0].id).toBe('refreshed')
-    })
-  })
-
-  // ──────────────────────── Merge items + presets ────────────────────────
-
-  it('merges items and presets from registry', async () => {
-    const items = [makeItem({ id: 'item-1', type: 'dashboard' })]
-    const presets = [makeItem({ id: 'preset-1', type: 'card-preset' })]
-    seedCache(items, presets)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems.length).toBe(2)
-    const ids = result.current.allItems.map(i => i.id)
-    expect(ids).toContain('item-1')
-    expect(ids).toContain('preset-1')
-  })
-
-  // ──────────────────────── reconcileImplementedCards ────────────────────────
-
-  it('promotes help-wanted item to available when card type is registered', async () => {
-    const items = [
-      makeItem({
-        id: 'cncf-karmada',
-        status: 'help-wanted',
-        tags: ['cncf', 'help-wanted'],
-      }),
-    ]
-    mockIsCardTypeRegistered.mockImplementation((t: string) => t === 'karmada_status')
-    seedCache(items)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems[0].status).toBe('available')
-    expect(result.current.allItems[0].tags).not.toContain('help-wanted')
-  })
-
-  it('maps cncf-flux to flux_status during reconcile', async () => {
-    const items = [
-      makeItem({
-        id: 'cncf-flux',
-        status: 'help-wanted',
-        tags: ['cncf', 'help-wanted'],
-      }),
-    ]
-    mockIsCardTypeRegistered.mockImplementation((t: string) => t === 'flux_status')
-    seedCache(items)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-
-    expect(result.current.allItems[0].status).toBe('available')
-    expect(result.current.allItems[0].tags).not.toContain('help-wanted')
-  })
-
-  it('does not promote help-wanted item when card type is NOT registered', async () => {
-    const items = [
-      makeItem({
-        id: 'cncf-karmada',
-        status: 'help-wanted',
-        tags: ['cncf', 'help-wanted'],
-      }),
-    ]
-    mockIsCardTypeRegistered.mockReturnValue(false)
-    seedCache(items)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems[0].status).toBe('help-wanted')
-    expect(result.current.allItems[0].tags).toContain('help-wanted')
-  })
-
-  it('leaves already-available items unchanged during reconcile', async () => {
-    const items = [
-      makeItem({
-        id: 'cncf-karmada',
-        status: 'available',
-        tags: ['cncf'],
-      }),
-    ]
-    seedCache(items)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems[0].status).toBe('available')
-    // Should NOT have called isCardTypeRegistered for already-available items
-    expect(mockIsCardTypeRegistered).not.toHaveBeenCalled()
-  })
-
-  it('does not reconcile help-wanted items that have no MARKETPLACE_TO_CARD_TYPE mapping', async () => {
-    const items = [
-      makeItem({
-        id: 'unknown-card-id',
-        status: 'help-wanted',
-        tags: ['help-wanted'],
-      }),
-    ]
-    seedCache(items)
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-    expect(result.current.allItems[0].status).toBe('help-wanted')
-  })
-
-  // ──────────────────────── Filtering ────────────────────────
-
-  it('filters items by search query (name)', async () => {
-    seedCache([
-      makeItem({ id: 'a', name: 'Monitoring Dashboard' }),
-      makeItem({ id: 'b', name: 'Security Scanner' }),
-    ])
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-
-    act(() => { result.current.setSearchQuery('monitor') })
-    expect(result.current.items.length).toBe(1)
-    expect(result.current.items[0].id).toBe('a')
-  })
-
-  it('filters items by search query (description)', async () => {
-    seedCache([
-      makeItem({ id: 'a', name: 'Thing', description: 'Monitors cluster health' }),
-      makeItem({ id: 'b', name: 'Other', description: 'Deploys applications' }),
-    ])
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-
-    act(() => { result.current.setSearchQuery('deploy') })
-    expect(result.current.items.length).toBe(1)
-    expect(result.current.items[0].id).toBe('b')
-  })
-
-  it('filters items by tag', async () => {
-    seedCache([
-      makeItem({ id: 'a', tags: ['monitoring', 'cncf'] }),
-      makeItem({ id: 'b', tags: ['security'] }),
-    ])
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-
-    act(() => { result.current.setSelectedTag('security') })
-    expect(result.current.items.length).toBe(1)
-    expect(result.current.items[0].id).toBe('b')
-  })
-
-  it('filters items by type', async () => {
-    seedCache([
-      makeItem({ id: 'a', type: 'dashboard' }),
-      makeItem({ id: 'b', type: 'theme' }),
-      makeItem({ id: 'c', type: 'card-preset' }),
-    ])
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-
-    act(() => { result.current.setSelectedType('theme') })
-    expect(result.current.items.length).toBe(1)
-    expect(result.current.items[0].id).toBe('b')
-  })
-
-  it('filters items by help-wanted status', async () => {
-    seedCache([
-      makeItem({ id: 'a', status: 'help-wanted' }),
-      makeItem({ id: 'b', status: 'available' }),
-      makeItem({ id: 'c' }), // no status — defaults to available
-    ])
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-
-    act(() => { result.current.setShowHelpWanted(true) })
-    expect(result.current.items.length).toBe(1)
-    expect(result.current.items[0].id).toBe('a')
-  })
-
-  it('combines multiple filters', async () => {
-    seedCache([
-      makeItem({ id: 'match', name: 'Monitoring', type: 'dashboard', tags: ['monitoring'] }),
-      makeItem({ id: 'wrong-type', name: 'Monitoring', type: 'theme', tags: ['monitoring'] }),
-      makeItem({ id: 'wrong-name', name: 'Security', type: 'dashboard', tags: ['monitoring'] }),
-    ])
-
-    const { result } = renderHook(() => useMarketplace())
-
-    await waitFor(() => {
-      expect(result.current.isLoading).toBe(false)
-    })
-
-    act(() => {
-      result.current.setSearchQuery('monitoring')
-      result.current.setSelectedType('dashboard')
-      result.current.setSelectedTag('monitoring')
-    })
-    expect(result.current.items.length).toBe(1)
-    expect(result.current.items[0].id).toBe('match')
-  })
-
-  // ──────────────────────── Tags / typeCounts / CNCF stats ────────────────────────
+  // ──────────────────────── Stats, installs, removes, localStorage ────────────────────────
 
   it('computes allTags as sorted unique set', async () => {
     seedCache([
@@ -991,165 +605,5 @@ describe('useMarketplace', () => {
 
     act(() => { result.current.setSearchQuery('GPU DASHBOARD') })
     expect(result.current.items.length).toBe(1)
-  })
-})
-
-// ---------------------------------------------------------------------------
-// Tests — useAuthorProfile
-// ---------------------------------------------------------------------------
-
-describe('useAuthorProfile', () => {
-  const AUTHOR_CACHE_PREFIX = 'kc-author-'
-
-  beforeEach(() => {
-    vi.clearAllMocks()
-    localStorage.clear()
-    vi.spyOn(globalThis, 'fetch').mockRejectedValue(new Error('not available'))
-  })
-
-  afterEach(() => {
-    vi.restoreAllMocks()
-  })
-
-  it('returns initial state when disabled', () => {
-    const { result } = renderHook(() => useAuthorProfile('testuser', false))
-    expect(result.current.loading).toBe(false)
-    expect(result.current.consolePRs).toBe(0)
-    expect(result.current.marketplacePRs).toBe(0)
-    expect(result.current.coins).toBe(0)
-  })
-
-  it('returns initial state when no handle', () => {
-    const { result } = renderHook(() => useAuthorProfile(undefined, true))
-    expect(result.current.loading).toBe(false)
-    expect(result.current.coins).toBe(0)
-  })
-
-  it('fetches PR counts from GitHub when enabled', async () => {
-    vi.mocked(globalThis.fetch)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 5 }),
-      } as Response)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 3 }),
-      } as Response)
-
-    const { result } = renderHook(() => useAuthorProfile('octocat', true))
-
-    await waitFor(() => {
-      expect(result.current.loading).toBe(false)
-      expect(result.current.consolePRs).toBe(5)
-    })
-    expect(result.current.marketplacePRs).toBe(3)
-    const COINS_PER_PR = 100
-    expect(result.current.coins).toBe((5 + 3) * COINS_PER_PR)
-  })
-
-  it('loads from valid cache without fetching', async () => {
-    const cached = {
-      consolePRs: 10,
-      marketplacePRs: 2,
-      fetchedAt: Date.now(),
-    }
-    localStorage.setItem(`${AUTHOR_CACHE_PREFIX}testuser`, JSON.stringify(cached))
-
-    const { result } = renderHook(() => useAuthorProfile('testuser', true))
-
-    await waitFor(() => {
-      expect(result.current.consolePRs).toBe(10)
-    })
-    expect(result.current.marketplacePRs).toBe(2)
-    const COINS_PER_PR = 100
-    expect(result.current.coins).toBe(12 * COINS_PER_PR)
-    // No fetch should have been called
-    expect(globalThis.fetch).not.toHaveBeenCalled()
-  })
-
-  it('ignores expired author cache', async () => {
-    const TWENTY_FIVE_HOURS_AGO = Date.now() - 25 * 60 * 60 * 1000
-    const cached = {
-      consolePRs: 10,
-      marketplacePRs: 2,
-      fetchedAt: TWENTY_FIVE_HOURS_AGO,
-    }
-    localStorage.setItem(`${AUTHOR_CACHE_PREFIX}staleuser`, JSON.stringify(cached))
-
-    vi.mocked(globalThis.fetch)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 20 }),
-      } as Response)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 5 }),
-      } as Response)
-
-    const { result } = renderHook(() => useAuthorProfile('staleuser', true))
-
-    await waitFor(() => {
-      expect(result.current.consolePRs).toBe(20)
-    })
-    expect(result.current.marketplacePRs).toBe(5)
-  })
-
-  it('returns 0 for PR counts when GitHub API fails', async () => {
-    vi.mocked(globalThis.fetch)
-      .mockResolvedValueOnce({ ok: false, status: 403 } as Response)
-      .mockResolvedValueOnce({ ok: false, status: 403 } as Response)
-
-    const { result } = renderHook(() => useAuthorProfile('ratelimited', true))
-
-    await waitFor(() => {
-      expect(result.current.loading).toBe(false)
-    })
-    expect(result.current.consolePRs).toBe(0)
-    expect(result.current.marketplacePRs).toBe(0)
-    expect(result.current.coins).toBe(0)
-  })
-
-  it('caches fetched results in localStorage', async () => {
-    vi.mocked(globalThis.fetch)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 7 }),
-      } as Response)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 1 }),
-      } as Response)
-
-    const { result } = renderHook(() => useAuthorProfile('cachetest', true))
-
-    await waitFor(() => {
-      expect(result.current.loading).toBe(false)
-      expect(result.current.consolePRs).toBe(7)
-    })
-
-    const stored = JSON.parse(localStorage.getItem(`${AUTHOR_CACHE_PREFIX}cachetest`)!)
-    expect(stored.consolePRs).toBe(7)
-    expect(stored.marketplacePRs).toBe(1)
-    expect(stored.fetchedAt).toBeDefined()
-  })
-
-  it('handles malformed author cache gracefully', async () => {
-    localStorage.setItem(`${AUTHOR_CACHE_PREFIX}badcache`, '<<not json>>')
-
-    vi.mocked(globalThis.fetch)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 2 }),
-      } as Response)
-      .mockResolvedValueOnce({
-        ok: true,
-        json: () => Promise.resolve({ total_count: 1 }),
-      } as Response)
-
-    const { result } = renderHook(() => useAuthorProfile('badcache', true))
-
-    await waitFor(() => {
-      expect(result.current.consolePRs).toBe(2)
-    })
   })
 })
